Safari-609.1.11.1/LayoutTests/js/regexp-find-first-asserted-expected.txt

Tests some regular expressions that were doing the wrong thing with the "find first asserted" optimization.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS /.*<body>(.*)<\/body>.*/.exec("foo<body>bar</body>baz").toString() is "foo<body>bar</body>baz,bar"
PASS /\s*<!--([sS]*)\/\/\s*-->\s*/.exec("<!--// -->").toString() is "<!--// -->,"

PASS successfullyParsed is true

TEST COMPLETE
